+++ src/main/java/healthcalc/adapter.java
@@ -12,7 +12,7 @@ public class adapter implements HealthHospital{
     public float pesoIdeal(float altura, char genero) throws Exception {
         // pasar a metros 
         // y pasar a gramos
-        return  adaptado.idealWeight((int) (altura*100), genero)*1000;
+        return  (float) (adaptado.idealWeight((int) (altura*100), genero)*1000);
     }
 
     @Override
@@ -20,7 +20,7 @@ public class adapter implements HealthHospital{
         // pasar a metros y gramos
        
         // y pasar otra vez a lo que tenia que ser
-        return adaptado.basalMetabolicRate(altura*100, edad, genero, genero)*1000;
+        return (float) (adaptado.basalMetabolicRate(altura*100, edad, genero, genero)*1000);
     }
     
 }
